==> notebooks/00_NST_DataPrep.ipynb <==
# Preliminaries and setup

1. Install and/or load required packages.
1. Load source data for *Challenge #2* provided by Piper project team.

## Install/Load required packages

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import re


from google.colab import drive
drive.mount('/content/drive')

## Load Raw Data

Load in the various **Piper $\nabla$** challenge data sets:

# Data files

urlC1a = ('https://drive.google.com/file/d/1Bl8O9FAYVHtIK4cdvqHAymeLFZUV9DZ0/view', 'Ch1 small')
urlC1b = ('https://drive.google.com/file/d/1BitVVnJweFfP38i-5p8FdsrN8Rr1DYh7/view', 'Ch1 medium')
urlC1c = ('https://drive.google.com/file/d/1iR447fRNgxIHkkE8oErv7ke3IDhGulIG/view', 'Ch1 large')

urlC2a = ('https://drive.google.com/file/d/12FPSxVnxzmvkxnS3nMD3MXKKIwxSytFf/view', 'Ch2 small')
urlC2c = ('https://drive.google.com/file/d/123oxrOJcUd2FeDmR6QBlznbTdwkcYpQk/view', 'Ch2 large')


Load data for *Challenge #2* (uncomment desired data set size): 

def read_from_url(url):
    return 'https://drive.google.com/uc?id='+url.split('/')[-2]

# read in C2 large file

# url = urlC2c[0]
# offline_tweets_df = pd.read_csv(read_from_url(url))

# read in C2 small file

# url = urlC2a[0]
url = urlC2c[0]
offline_tweets_df = pd.read_csv(read_from_url(url), index_col=0)

display(offline_tweets_df)


# Correct swapped `id` and `created_at` column names
offline_tweets_df.rename(columns={'id':'created_at', 'created_at':'id'}, inplace=True)

offline_tweets_df.head()

# Data Parsing

1. Initial data cleaning
1. Parsing mentions, retweets, hastags, and web links

## Intial Data Cleaning

# replace Nans in user_description with empty strings
offline_tweets_df.user_description.fillna('', inplace=True)
print('Nans replaced')

# convert comments ("text") from byte encoding to utf8 
import ast
offline_tweets_df['text2']=offline_tweets_df['text'].apply(lambda x: ast.literal_eval(x).decode('utf-8'))
#offline_tweets_df['userdescr']=offline_tweets_df['user_description'].apply(lambda x: ast.literal_eval(x).decode('utf-8'))
display(offline_tweets_df[['user_description','text','text2']])

# convert user_description from byte encoding to utf8 

# not all rows are byte-encoded so need to create a mask:
is_byte = (offline_tweets_df.user_description.str.startswith("b'") | offline_tweets_df.user_description.str.startswith('b"'))

# vector of clean utf8 encoded user_descriptions
clean_ud = offline_tweets_df.loc[is_byte,'user_description'].apply(lambda x: ast.literal_eval(x).decode('utf-8'))
print(len(clean_ud))

clean_ud.head()

offline_tweets_df.loc[is_byte, 'user_description'] = clean_ud
display(offline_tweets_df[['user_description','text','text2']])


from bs4 import BeautifulSoup
# from nltk.tokenize import WordPunctTokenizer

# tok = WordPunctTokenizer()

# pat1 = r'@[A-Za-z0-9]+'
# pat2 = r'https?://[A-Za-z0-9./]+'
# combined_pat = r'|'.join((pat1, pat2))

# Convert ascii codes and line breaks
def tweet_cleaner(text):
    soup = BeautifulSoup(text, 'lxml')
    souped = soup.get_text()
    return re.sub(r"\n", ' ', souped)

offline_tweets_df['text3'] = offline_tweets_df['text2'].map(lambda x: tweet_cleaner(x))
offline_tweets_df['user_descr'] = offline_tweets_df['user_description'].map(lambda x: tweet_cleaner(x))

display(offline_tweets_df[['text2','text3']][offline_tweets_df.text2!=offline_tweets_df.text3])


## Parsing Mentions, Retweets, Hastags, and Links

# Functions to identify retweets, mentions, hashtags, and links

def find_retweeted(tweet):
  '''This function will extract the twitter handles of retweed people'''
  return re.findall('(?<=RT\s)(@[A-Za-z]+[A-Za-z0-9-_]+)', tweet)

def find_mentioned(tweet):
  '''This function will extract the twitter handles of people mentioned in the tweet'''
  return re.findall('(?<!RT\s)(@[A-Za-z]+[A-Za-z0-9-_]+)', tweet)  

def find_hashtags(tweet):
  '''This function will extract hashtags'''
  return re.findall('(#[A-Za-z]+[A-Za-z0-9-_]+)', tweet)  

def find_links(tweet):
  '''This function will extract url links'''
  http_pattern = r'https?://[A-Za-z0-9./]+'
  bitly_pattern = r'bit.ly/\S+'
  pattern = r'|'.join((http_pattern, bitly_pattern))
  return re.findall(pattern, tweet)


# make new columns for retweeted usernames, mentioned usernames and hashtags

offline_tweets_df['is_retweet'] = offline_tweets_df['text3'].apply(lambda x: x[:2]=='RT')
offline_tweets_df['retweeted_from'] = offline_tweets_df.text3.apply(find_retweeted)
offline_tweets_df['mentioned'] = offline_tweets_df.text3.apply(find_mentioned)
offline_tweets_df['hashtags'] = offline_tweets_df.text3.apply(find_hashtags)
offline_tweets_df['links'] = offline_tweets_df.text3.apply(find_links)

display(offline_tweets_df[['is_retweet','retweeted_from','mentioned','hashtags','links']].head(10))


#Substitutions
Substitute http for links and @user for mentions in tweet text

pat1 = r'@[A-Za-z0-9]+'
pat2 = r'https?://[A-Za-z0-9./]+'

def preprocess(text, pat1repl='@user', pat2repl='http'):
    stripped = re.sub(pat1, pat1repl, text)
    stripped = re.sub(pat2, pat2repl, stripped)
    return stripped

text = offline_tweets_df['text2'].map(lambda x: preprocess(x))


# Save Prepared Dataframe

offline_tweets_df.to_pickle('/content/drive/MyDrive/Piper Gradient/Not-So-Twitterpated/cleaned_offline_tweets_df_large.pickle')

# Coding References

* [Topic Extraction from Tweets using LDA, by *Usen Osas*](https://medium.com/@osas.usen/topic-extraction-from-tweets-using-lda-a997e4eb0985)

* [Topic Modelling in Python: Unsupervised Machine Learning to Find Tweet Topics](https://ourcodingclub.github.io/tutorials/topic-modelling-python/)
